build(protocol): 抽出与 STM32 共享的 C99 协议核心 edge_proto
协议契约此前靠三处手抄维持:网关 GatewayApp.cpp 的匿名 namespace、
节点 cmd_service.c 的 switch 裸数字、fake_stm32.cpp 自抄的第三份组帧。
无人保证三者一致,结果 0x22 采样周期的单位在文档/两端代码/README 里
形成四处三种说法而长期无人发觉。

引入 protocol/ 作为唯一真相:纯 C99、无 malloc/stdio/平台头、不做 I/O,
两个仓库编译同一份源码。edge_proto.h 收纳 TYPE 字典、结果码、payload
布局、量纲、时序契约,并把单位写进标识符名字(EDGE_PERIOD_*_S)。

- 收帧 FSM 由两端实现归一,错误改为累计到 edge_parser_stats_t,
  不再像原网关那样直接写 std::cerr(协议层不该决定错误怎么呈现)
- vectors/ 金标准向量两端共读:文档 §4.3 四条 CRC 向量 + §7 全部八个
  完整帧示例(已与原实现逐字节交叉验证)+ 边界与错误路径
- MANIFEST.sha256 + check_proto_sync.sh 把「两端是否同步」变成
  机器可判定的事实,已验证篡改时确实返回非零

同时补齐构建基建:CMakePresets(dev/asan/tsan/release)、
cmake/CompilerWarnings.cmake、GoogleTest + CTest、.clang-format。

新警告集(-Wconversion -Wsign-conversion -Wshadow 等)暴露 20 处符号
转换问题,逐一修复而非降低标准。其中 Logger::log 是真 bug:snprintf
返回「本应写入的长度」,前缀超长时 buf + prefix_len 越界且
sizeof(buf) - prefix_len 下溢成天文数字。

tests: 24/24 通过
arm-none-eabi-gcc -mcpu=cortex-m3 -ffreestanding 编译通过(972B text, 0 bss)

+++ src/log/AsyncLogger.cpp
@@ -10,18 +10,18 @@
 namespace gateway {
 
     static ssize_t safe_write(int fd, const void* buf, size_t n) {
-        ssize_t total = 0;
+        size_t total = 0;                       // 用无符号计数,与 n 同类型,免去来回强转
         const char* p = static_cast<const char*>(buf);
-        while(total < static_cast<ssize_t>(n)) {
+        while(total < n) {
             ssize_t n_write = write(fd, p + total, n - total);
             if (n_write < 0) {
                 if( errno == EINTR) continue;
                 return -1;
             }
-            total += n_write;
+            total += static_cast<size_t>(n_write);   // 此处 n_write >= 0,转换安全
         }
 
-        return total;
+        return static_cast<ssize_t>(total);
     }
 
     AsyncLogger& AsyncLogger::instance() {
